renamed new python script to ui_tasks so it can be used more generally
